requirements.txt
fastapi==0.136.1
google_api_python_client==2.193.0
google_auth_oauthlib==1.3.1
protobuf==7.34.1
pydantic==2.13.3
python-dotenv==1.2.2
starlette==1.0.0